# Conflicts:
#	com.unity.render-pipelines.universal/Runtime/UniversalRenderPipelineCore.cs

+++ TestProjects/VisualEffectGraph_HDRP/Assets/AllTests/Editor/Tests/VFXTestCommon.cs
@@ -8,13 +8,43 @@ using System.Linq;
 using System.Collections.Generic;
 using UnityEngine.UIElements;
 using UnityEditor.Experimental.GraphView;
+using System.IO;
+using UnityEngine.Playables;
+using UnityEngine.Timeline;
 
 namespace UnityEditor.VFX.Test
 {
     class VFXTestCommon
     {
-        static readonly string tempBasePath = "Assets/TmpTests/";
+        public static readonly string tempBasePath = "Assets/TmpTests/";
         static readonly string tempFileFormat = tempBasePath + "vfx_{0}.vfx";
+        static readonly string tempFileFormatPlayable = tempBasePath + "vfx_{0}.playable";
+
+        public static VFXGraph CopyTemporaryGraph(string path)
+        {
+            var guid = System.Guid.NewGuid().ToString();
+            string tempFilePath = string.Format(tempFileFormat, guid);
+            System.IO.Directory.CreateDirectory(tempBasePath);
+            File.Copy(path, tempFilePath);
+
+            AssetDatabase.ImportAsset(tempFilePath);
+            var asset = AssetDatabase.LoadAssetAtPath<VisualEffectAsset>(tempFilePath);
+            VisualEffectResource resource = asset.GetResource();
+            var graph = resource.GetOrCreateGraph();
+            return graph;
+        }
+
+        public static TimelineAsset CopyTemporaryTimeline(string path)
+        {
+            var guid = System.Guid.NewGuid().ToString();
+            string tempFilePath = string.Format(tempFileFormatPlayable, guid);
+            System.IO.Directory.CreateDirectory(tempBasePath);
+            File.Copy(path, tempFilePath);
+
+            AssetDatabase.ImportAsset(tempFilePath);
+            var asset = AssetDatabase.LoadAssetAtPath<TimelineAsset>(tempFilePath);
+            return asset;
+        }
 
         public static VFXGraph MakeTemporaryGraph()
         {
@@ -30,14 +60,17 @@ namespace UnityEditor.VFX.Test
 
         public static void DeleteAllTemporaryGraph()
         {
-            foreach (string file in System.IO.Directory.GetFiles(tempBasePath))
+            if (Directory.Exists(tempBasePath))
             {
-                try
+                foreach (string file in System.IO.Directory.GetFiles(tempBasePath))
                 {
-                    AssetDatabase.DeleteAsset(file);
-                }
-                catch (System.Exception) // Don't stop if we fail to delete one asset
-                {
+                    try
+                    {
+                        AssetDatabase.DeleteAsset(file);
+                    }
+                    catch (System.Exception) // Don't stop if we fail to delete one asset
+                    {
+                    }
                 }
             }
         }
